Implemented the units; Community Spirit, Material Satisfaction, Relational Satisfaction and Money with UI
Satisfaction is divided in to MS (red) and RS (yellow). They are in a circular meter that is prefabed. More RS increases Community spirit and more MS decreases Community Spirit.

Added a script called Resource Manager that is attached to the Game Manager object. This script holds all the logic for the units.

Added test buttons that have a test script called Transaction which communicates value changes to the Resource Manager script.

Community Spirit slider is prefabed.

Money is a simple TextMesh (not prefabed).