Shop System and Updated ResourceManager and UI
Added new UI elements to the Community Spirit Meter and Created a Prefab for it in the Prefab UI folder. "CommunitySpiritMeter". Uses more similar logic to SatisfactionMeter.

Bug fixes with satisfaction meter.

Optimized ResourceManager code.

Added a shop prefab to the prefab UI folder. What u buy impacts resources in the resource manager and becomes visible in the player inventory. The shop buttons uses the Transaction script.

Added a Test Shop prefab to Prefab Triggers folder. Uses the new OpenShop script. Player presses interact while in the test shop area to open the shop. If player exits the area or presses interact again, the shop disappears.

Added a full screen shader that activates when material satisfaction is high. Its a raw image that the resource manger can use to turn on and off the alpha of the image when conditions are met. Prefabed in Prefabs UI folder.